Version 1.5.2 (Effective 19.03)
merged the function that draws the playing field and the function that
draws Karel to one draw function.
New variables "columns" and "rows" which represent the columns and rows
in Karel's world.
body onload="…" is now changed to "$(window).load(...", thereby
JavaScript and HTML is not mixed anymore.
Better commentation.
Code Cleaning.